backend/pyproject.toml
[project]
name = "ai-voice-chat-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "fastrtc[stt,vad]>=0.0.30",
    "openai>=1.0.0",
    "pocket-tts>=1.0.0",
    "numpy>=2.0.0",
    "scipy>=1.14.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]
